# api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: QuantumLab API
  version: v1
  summary: REST API платформы QuantumLab
  description: |
    Единый контракт для Web UI, CLI и внешних интеграций.

    Правила:
      * версионирование в пути (`/api/v1`), обратная совместимость в пределах мажорной версии;
      * ошибки возвращаются в формате RFC 9457 (`application/problem+json`) с кодом из
        таксономии `quantumlab.errors.ErrorCode` и локализованным диагнозом;
      * тяжёлые научные данные не отдаются телом ответа — только ссылки на артефакты;
      * прогресс задания доступен опросом `GET /jobs/{id}` и потоком
        `GET /jobs/{id}/events` (SSE).
  license:
    name: Apache-2.0
servers:
  - url: /api/v1
    description: Относительный префикс — API развёрнут вместе с UI
...
security:
  - bearerAuth: []
paths:
  /auth/token:
    post:
      tags: [auth]
      operationId: createToken
      summary: Получить токен доступа
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Токен выдан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Token' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /projects:
    get:
      tags: [projects]
      operationId: listProjects
      summary: Список проектов пользователя
      parameters:
        - { $ref: '#/components/parameters/AcceptLanguage' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Страница проектов
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Project' }
                  total: { type: integer, minimum: 0 }
    post:
      tags: [projects]
      operationId: createProject
      summary: Создать проект
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreate' }
      responses:
        '201':
          description: Проект создан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
  /projects/{projectId}/molecules:
    get:
      tags: [molecules]
      operationId: listMolecules
      summary: Молекулы проекта
      parameters:
        - { $ref: '#/components/parameters/ProjectId' }
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Список структур
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Molecule' }
    post:
      tags: [molecules]
      operationId: createMolecule
      summary: Добавить структуру (из файла или по SMILES/InChI)
      parameters:
        - { $ref: '#/components/parameters/ProjectId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MoleculeCreate' }
      responses:
        '201':
          description: Структура создана; в ответе — результат проверок
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Molecule' }
        '422': { $ref: '#/components/responses/Problem' }
  /molecules/{moleculeId}:
    get:
      tags: [molecules]
      operationId: getMolecule
      summary: Получить структуру
      parameters:
        - { $ref: '#/components/parameters/MoleculeId' }
        - { $ref: '#/components/parameters/AcceptLanguage' }
      responses:
        '200':
          description: Структура
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Molecule' }
        '404': { $ref: '#/components/responses/NotFound' }
  /molecules/{moleculeId}/validate:
    post:
      tags: [molecules]
      operationId: validateMolecule
      summary: Проверить структуру (валентности, заряд/мультиплетность, контакты)
      parameters:
        - { $ref: '#/components/parameters/MoleculeId' }
        - { $ref: '#/components/parameters/AcceptLanguage' }
      responses:
        '200':
          description: Отчёт проверок
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StructureReport' }
  /calculations/plan:
    post:
      tags: [calculations]
      operationId: planCalculation
      summary: Развернуть профиль точности в полную спецификацию с обоснованиями
      description: |
        Ключевой эндпоинт «рекомендуемых настроек» (§8 ТЗ): на вход — задача,
        профиль и идентификатор структуры; на выход — полная спецификация и
        список решений с текстовыми обоснованиями для показа пользователю.
      parameters:
        - { $ref: '#/components/parameters/AcceptLanguage' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlanRequest' }
      responses:
        '200':
          description: Спецификация и обоснования
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanResponse' }
  /jobs:
    get:
      tags: [calculations]
      operationId: listJobs
      summary: Очередь и история заданий
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/JobStatus' }
        - name: projectId
          in: query
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Страница заданий
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Job' }
                  total: { type: integer, minimum: 0 }
    post:
      tags: [calculations]
      operationId: submitJob
      summary: Отправить расчёт
      parameters:
        - { $ref: '#/components/parameters/AcceptLanguage' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobCreate' }
      responses:
        '202':
          description: Задание принято в очередь
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '422': { $ref: '#/components/responses/Problem' }
  /jobs/{jobId}:
    get:
      tags: [calculations]
      operationId: getJob
      summary: Состояние задания (статус, прогресс, ETA, ресурсы)
      parameters:
        - { $ref: '#/components/parameters/JobId' }
        - { $ref: '#/components/parameters/AcceptLanguage' }
      responses:
        '200':
          description: Задание
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '404': { $ref: '#/components/responses/NotFound' }
  /jobs/{jobId}/cancel:
    post:
      tags: [calculations]
      operationId: cancelJob
      summary: Остановить задание
      parameters:
        - { $ref: '#/components/parameters/JobId' }
      responses:
        '200':
          description: Задание остановлено
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '409': { $ref: '#/components/responses/Problem' }
  /jobs/{jobId}/resume:
    post:
      tags: [calculations]
      operationId: resumeJob
      summary: Продолжить расчёт с контрольной точки
      parameters:
        - { $ref: '#/components/parameters/JobId' }
        - { $ref: '#/components/parameters/AcceptLanguage' }
      responses:
        '200':
          description: Расчёт продолжается
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '409':
          description: Нет контрольной точки (`job.not_resumable`)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /jobs/{jobId}/retry:
    post:
      tags: [calculations]
      operationId: retryJob
      summary: Повторить задание (attempt += 1)
      parameters:
        - { $ref: '#/components/parameters/JobId' }
      responses:
        '202':
          description: Задание снова в очереди
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
  /jobs/{jobId}/events:
    get:
      tags: [calculations]
      operationId: streamJobEvents
      summary: Поток событий задания (Server-Sent Events)
      parameters:
        - { $ref: '#/components/parameters/JobId' }
      responses:
        '200':
          description: Поток событий прогресса и смены состояний
          content:
            text/event-stream:
              schema: { type: string }
  /jobs/{jobId}/logs:
    get:
      tags: [calculations]
      operationId: getJobLogs
      summary: Журнал расчёта
      parameters:
        - { $ref: '#/components/parameters/JobId' }
        - name: tail
          in: query
          schema: { type: integer, minimum: 1, default: 500 }
      responses:
        '200':
          description: Хвост журнала
          content:
            text/plain:
              schema: { type: string }
  /jobs/{jobId}/result:
    get:
      tags: [results]
      operationId: getResult
      summary: Результат расчёта (скаляры, проверки качества, ссылки на артефакты)
      parameters:
        - { $ref: '#/components/parameters/JobId' }
        - { $ref: '#/components/parameters/AcceptLanguage' }
      responses:
        '200':
          description: Результат
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CalculationResult' }
        '404': { $ref: '#/components/responses/NotFound' }
  /jobs/{jobId}/report:
    get:
      tags: [results]
      operationId: getReport
      summary: Отчёт по расчёту (§39 ТЗ)
      parameters:
        - { $ref: '#/components/parameters/JobId' }
        - name: format
          in: query
          schema: { type: string, enum: [html, pdf, json], default: html }
      responses:
        '200':
          description: Файл отчёта
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
  /jobs/{jobId}/result/export:
    post:
      tags: [results]
      operationId: exportResults
      summary: Пакетный экспорт результатов (CSV / JSON / Parquet)
      parameters:
        - { $ref: '#/components/parameters/JobId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExportRequest' }
      responses:
        '202':
          description: Экспорт поставлен в очередь
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportTask' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      description: Язык сообщений и диагнозов (по умолчанию `ru`)
      schema: { type: string, enum: [ru, en], default: ru }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
...
    Unauthorized:
      description: Требуется аутентификация
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Token:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token: { type: string }
        token_type: { type: string, enum: [Bearer] }
        expires_in: { type: integer, description: секунд }
...
    Project:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        role: { type: string, enum: [owner, editor, viewer] }
        createdAt: { type: string, format: date-time }
...